Replacing acme_compact with a simpler certtool.py.
